# api-document.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "API Document",
        "description": "Mota API Service",
        "license": {
            "name": "API Licence",
            "url": "http://domain.vn/license"
        },
        "version": "v1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "server test"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
        "/movie/{id}/update": {
            "put": {
                "tags": [
                    "movie-controller"
                ],
                "operationId": "updateMovie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMovieRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/MovieDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/user/create": {
            "post": {
                "tags": [
                    "User Controller"
                ],
                "summary": "Add user",
                "description": "Api create new user",
                "operationId": "createUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegisterDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegisterDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movie/create": {
            "post": {
                "tags": [
                    "movie-controller"
                ],
                "operationId": "createMovie",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateMovieRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/MovieDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResLoginDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/search/movies": {
            "get": {
                "tags": [
                    "search-controller"
                ],
                "operationId": "searchMovies",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "genreIds",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/MovieSearchResponseDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/recent/user/{id}": {
            "get": {
                "tags": [
                    "movie-controller"
                ],
                "operationId": "getMovieHistoryByUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WatchHistoryProjection"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/movies/{id}": {
            "get": {
                "tags": [
                    "movie-controller"
                ],
                "operationId": "getMovieByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/MovieDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/movies/type": {
            "get": {
                "tags": [
                    "movie-controller"
                ],
                "operationId": "getAllMoviesByType",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "single",
                                "series"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MovieDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/genre/{id}": {
            "get": {
                "tags": [
                    "movie-controller"
                ],
                "operationId": "getMovieByGenreID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MovieDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/actor/{id}": {
            "get": {
                "tags": [
                    "movie-controller"
                ],
                "operationId": "getMovieByActorID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MovieDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/actor/{id}": {
            "get": {
                "tags": [
                    "actor-controller"
                ],
                "operationId": "getActorById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActorDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/movie/{id}/delete": {
            "delete": {
                "tags": [
                    "movie-controller"
                ],
                "operationId": "deleteMovie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ActorRequest": {
                "type": "object",
                "properties": {
                    "actorId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "roleName": {
                        "type": "string"
                    }
                }
            },
            "EpisodeRequest": {
                "type": "object",
                "properties": {
                    "episodeNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "title": {
                        "type": "string"
                    },
                    "videoUrl": {
                        "type": "string"
                    },
                    "duration": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "UpdateMovieRequest": {
                "required": [
                    "description",
                    "title"
                ],
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "duration": {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int32"
                    },
                    "country": {
                        "type": "string"
                    },
                    "language": {
                        "type": "string"
                    },
                    "ageLimit": {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int32"
                    },
                    "trailerUrl": {
                        "type": "string"
                    },
                    "posterUrl": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "single",
                            "series"
                        ]
                    },
                    "genreIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    "actors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ActorRequest"
                        }
                    },
                    "episodes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EpisodeRequest"
                        }
                    }
                }
            },
...
            "MovieDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "duration": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "country": {
                        "type": "string"
                    },
                    "language": {
                        "type": "string"
                    },
                    "ageLimit": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "trailerUrl": {
                        "type": "string"
                    },
                    "posterUrl": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "single",
                            "series"
                        ]
                    },
                    "releaseDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "actors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "genres": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "episodes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EpisodeDTO"
                        }
                    },
                    "continueWatching": {
                        "$ref": "#/components/schemas/WatchHistoryDTO"
                    }
                }
            },
...
            "ActorDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "birthDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "nationality": {
                        "type": "string"
                    },
                    "bio": {
                        "type": "string"
                    },
                    "avatarUrl": {
                        "type": "string"
                    }
                }
            },
            "CreateUserRequest": {
                "required": [
                    "email",
                    "fullName",
                    "password",
                    "username"
                ],
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "RegisterDTO": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    }
                }
            },
            "CreateMovieRequest": {
                "required": [
                    "title"
                ],
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "duration": {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int32"
                    },
                    "country": {
                        "type": "string"
                    },
                    "language": {
                        "type": "string"
                    },
                    "ageLimit": {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int32"
                    },
                    "trailerUrl": {
                        "type": "string"
                    },
                    "posterUrl": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "single",
                            "series"
                        ]
                    },
                    "genreIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    "actors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ActorRequest"
                        }
                    },
                    "episodes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EpisodeRequest"
                        }
                    }
                }
            },
            "LoginDTO": {
                "required": [
                    "password",
                    "username"
                ],
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "ResLoginDTO": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    }
                }
            },
...
            "MovieSearchResponseDto": {
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MovieSearchItemDto"
                        }
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "total": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "WatchHistoryProjection": {
                "type": "object",
                "properties": {
                    "episodeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "lastWatchedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "posterUrl": {
                        "type": "string"
                    },
                    "watchedSeconds": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "episodeNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "completed": {
                        "type": "boolean"
                    },
                    "movieId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}